bugfix: Properly set the _malloced field of PyInterpreterState
When creating a new sub-interpreter, the exact address that should be
freed by `PyMem_RawFree` (when the interpreter is deleted) is stored in
the `_malloced` field of the allocated `PyInterpreterState` struct.
However, the content of this allocated struct is overridden by `memcpy`
so the content of the `_malloced` field is lost and replaced by NULL.
This commit solves this by storing the `_malloced` field in a temporary
variable and restors it after the memcpy.